* FIX: SI::Timing IPC

+++ src/arch/southern-islands/timing/Timing.cc
@@ -1005,7 +1005,8 @@ void Timing::DumpReport() const {
     long long coalesced_writes =
         compute_unit->getLdsModule()->num_coalesced_writes;
     instructions_per_cycle =
-        getCycle() ? (double)(compute_unit->num_total_instructions / getCycle())
+        getCycle() ? ((double)compute_unit->num_total_instructions /
+                      (double)getCycle())
                    : 0.0;
 
     // Report statistics for each compute unit
